Change default behavior when including the gem
Before we were permitting all the parameters and forcing to whitelist all
the attributes. This make harder to upgrade to strong parameters because
you can't slowly change the models to not use the attribute macros.

Since now we are fallbacking to strong parameters if the macro is not
defined on the model we should not permit all the parameters neither
call the macro on ActiveRecord::Base.

Models with macros will use the mass assignment sanitization and models
without the macros will raise if parameters are not permitted.